+++ src/army/generic/abilities.ts
@@ -1,0 +1,14 @@
+import { TURN_ONE_START_OF_HERO_PHASE } from 'types/phases'
+
+// This file is useful when storing abilities for units that we'd like to re-use
+// E.G. all mercenaries share the "Disruptive Presence" effect, so we'll store it here.
+
+const GenericAbilities = {
+  DisruptivePresence: {
+    name: `Disruptive Presence`,
+    desc: `If your army includes any MERCENARY units, at the start of your hero phase in the first battle round, you do not receive 1 command point.`,
+    when: [TURN_ONE_START_OF_HERO_PHASE],
+  },
+}
+
+export default GenericAbilities
