+++ src/army/mega_gargant_mercenaries/abilities.ts
