+++ src/army/mega_gargant_mercenaries/index.ts
@@ -1,0 +1,6 @@
+import Abilities from './abilities'
+import { Units } from './units'
+
+const MegaGargantMercenariesArmy = { Abilities, Units }
+
+export default MegaGargantMercenariesArmy
