+++ src/army/mercenary_companies/index.ts
@@ -1,9 +1,6 @@
 import Abilities from './abilities'
 import { Units } from './units'
 
-const MercenaryCompaniesArmy = {
-  Abilities,
-  Units,
-}
+const MercenaryCompaniesArmy = { Abilities, Units }
 
 export default MercenaryCompaniesArmy
