--- src/army/sons_of_behemat/endless_spells.ts
@@ -1,18 +1,0 @@
-import { TEndlessSpells } from 'types/army'
-import { START_OF_ROUND } from 'types/phases'
-
-// Endless spells go here
-const EndlessSpells: TEndlessSpells = [
-  {
-    name: ``,
-    effects: [
-      {
-        name: ``,
-        desc: ``,
-        when: [START_OF_ROUND],
-      },
-    ],
-  },
-]
-
-export default EndlessSpells
