+++ src/army/sons_of_behemat/index.ts
@@ -1,24 +1,14 @@
 import Abilities from './abilities'
 import Allegiances from './allegiances'
 import Artifacts from './artifacts'
-import EndlessSpells from './endless_spells'
-import Scenery from './scenery'
-import Spells from './spells'
 import Traits from './traits'
-import { AlliedUnits, Battalions, Units } from './units'
-
-// TODO: Handle Mercenary aspect
+import { Units } from './units'
 
 const SonsOfBehematArmy = {
   Abilities,
   Allegiances,
   AllegianceType: 'Tribes',
-  AlliedUnits,
   Artifacts,
-  Battalions,
-  EndlessSpells,
-  Scenery,
-  Spells,
   Traits,
   Units,
 }
