--- src/army/sons_of_behemat/scenery.ts
@@ -1,16 +1,0 @@
-import { TScenery } from 'types/army'
-
-const Scenery: TScenery = [
-  {
-    name: ``,
-    effects: [
-      {
-        name: ``,
-        desc: ``,
-        when: [],
-      },
-    ],
-  },
-]
-
-export default Scenery
