--- src/army/sons_of_behemat/spells.ts
@@ -1,19 +1,0 @@
-import { TSpells } from 'types/army'
-import { HERO_PHASE } from 'types/phases'
-
-// Spells, Prayers, etc. go here when they're part of the allegiance
-// If they're on a warscroll, put them on the unit in units.ts instead
-const Spells: TSpells = [
-  {
-    name: ``,
-    effects: [
-      {
-        name: ``,
-        desc: ``,
-        when: [HERO_PHASE],
-      },
-    ],
-  },
-]
-
-export default Spells
